Refactor browser extension Rust bridge (#2011)
